refactor: anime/detail 변경된 파일 구조 적용
