+++ src/features/animes/routes/Detail/Hero/style.ts
