+++ src/features/animes/routes/Detail/PlotAndInfo/style.ts
