+++ src/features/animes/routes/Detail/Ratings/style.ts
