+++ src/features/animes/routes/Detail/Reviews/style.ts
@@ -1,6 +1,6 @@
 import styled from "@emotion/styled";
 
-import { SectionContainer } from "./Section.style";
+import { SectionContainer } from "../Section.style";
 
 export const Section = styled(SectionContainer)`
   padding-bottom: 200px;
